Добавлена логика в файлах

+++ smtuIdle/models.py
@@ -38,7 +38,7 @@ class Purchase(BaseModel):
     AuctionDate = DateField(null=True,  default="Нет данных", verbose_name="Дата аукциона")
     notification_link = CharField(null=True,verbose_name="Извещение о закупке", max_length=255)
     quantity_units = IntegerField(null=True,verbose_name="Количество единиц")
-    nmck_per_unit = IntegerField(null=True,verbose_name="НМЦК за единицу")
+    nmck_per_unit = FloatField(null=True,verbose_name="НМЦК за единицу")
     nmck_file = CharField(null=True,verbose_name="Файл НМЦК", max_length=255)
     protocol_file = CharField(null=True,verbose_name="Файл Протокола", max_length=255)
     # Добавленные поля
